# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Rivora Exchange API
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------
  # Users
  # ---------------------------------------------------------------------
...
  /investment-plans:
    get:
      operationId: listInvestmentPlans
      tags: [investment-plans]
      parameters:
        - name: activeOnly
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvestmentPlan'
    post:
      operationId: createInvestmentPlan
      tags: [investment-plans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvestmentPlanRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestmentPlan'
  /investment-plans/{planId}:
    patch:
      operationId: updateInvestmentPlan
      tags: [investment-plans]
      parameters:
        - name: planId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInvestmentPlanRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestmentPlan'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteInvestmentPlan
      tags: [investment-plans]
      parameters:
        - name: planId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------
  # Investments (user placing money into a plan)
  # ---------------------------------------------------------------------
...
  /deposit-requests:
    get:
      operationId: listDepositRequests
      tags: [deposit-requests]
      parameters:
        - name: userId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/RequestStatus' }
      responses:
        '200':
          description: Deposit requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DepositRequest'
    post:
      operationId: createDepositRequest
      tags: [deposit-requests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepositRequestBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositRequest'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /deposit-requests/{requestId}:
    patch:
      operationId: updateDepositRequest
      tags: [deposit-requests]
      parameters:
        - name: requestId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRequestStatusBody'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositRequest'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------
  # Withdrawal requests
  # ---------------------------------------------------------------------
...
components:
  schemas:
    UserRole:
      type: string
      enum: [user, admin]
    UserStatus:
      type: string
      enum: [active, inactive]
...
    RequestStatus:
      type: string
      enum: [pending, approved, rejected]
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    User:
      type: object
      required: [id, phone, fullName, role, status, balance, referralCode, hasReceivedWelcomeBonus, mustChangePassword, createdAt, updatedAt]
      properties:
        id: { type: string }
        phone: { type: string }
        fullName: { type: string }
        role: { $ref: '#/components/schemas/UserRole' }
        status: { $ref: '#/components/schemas/UserStatus' }
        balance: { type: number }
        referralCode: { type: string }
        referredBy: { type: string, nullable: true }
        hasReceivedWelcomeBonus: { type: boolean }
        bankName: { type: string, nullable: true }
        bankAccountNumber: { type: string, nullable: true }
        bankAccountName: { type: string, nullable: true }
        mustChangePassword: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user: { $ref: '#/components/schemas/User' }
        token: { type: string }
    RegisterRequest:
      type: object
      required: [phone, password, confirmPassword, fullName]
      properties:
        phone: { type: string, minLength: 10 }
        password: { type: string, minLength: 6 }
        confirmPassword: { type: string, minLength: 6 }
        fullName: { type: string, minLength: 1 }
        referralCode: { type: string, nullable: true }
    LoginRequest:
      type: object
      required: [phone, password]
      properties:
        phone: { type: string }
        password: { type: string }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 6 }
    InvestmentPlan:
      type: object
      required: [id, name, dailyRate, minAmount, durationDays, isActive, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        dailyRate: { type: number }
        minAmount: { type: number }
        durationDays: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateInvestmentPlanRequest:
      type: object
      required: [name, dailyRate, minAmount, durationDays]
      properties:
        name: { type: string }
        dailyRate: { type: number }
        minAmount: { type: number }
        durationDays: { type: integer }
        isActive: { type: boolean }
    UpdateInvestmentPlanRequest:
      type: object
      properties:
        name: { type: string }
        dailyRate: { type: number }
        minAmount: { type: number }
        durationDays: { type: integer }
        isActive: { type: boolean }
...
    DepositRequest:
      type: object
      required: [id, userId, amount, paymentMethod, status, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        amount: { type: number }
        paymentMethod: { type: string }
        proofUrl: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/RequestStatus' }
        createdAt: { type: string, format: date-time }
        reviewedAt: { type: string, format: date-time, nullable: true }
    CreateDepositRequestBody:
      type: object
      required: [amount, paymentMethod]
      properties:
        amount: { type: number }
        paymentMethod: { type: string }
        proofUrl: { type: string, nullable: true }
...
    UpdateRequestStatusBody:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/RequestStatus' }